+++ h5netcdf/core.py
@@ -615,7 +615,7 @@ class Group(Mapping):
         return group._create_child_group(keys[-1])
 
     def _create_child_variable(
-        self, name, dimensions, dtype, data, fillvalue, **kwargs
+        self, name, dimensions, dtype, data, fillvalue, chunks, **kwargs
     ):
         if name in self:
             raise ValueError(
@@ -662,6 +662,16 @@ class Group(Mapping):
         if shape != maxshape:
             kwargs["maxshape"] = maxshape
 
+        if isinstance(chunks, str):
+            if chunks == "h5py":
+                chunks = True
+            elif chunks == "h5netcdf":
+                chunks = _get_default_chunksizes(maxshape, dtype)
+            else:
+                raise ValueError(
+                    "got unrecognized string value %s for chunks argument" % chunks
+                )
+
         # Clear dummy HDF5 datasets with this name that were created for a
         # dimension scale without a corresponding variable.
         # Keep the references, to re-attach later
@@ -672,15 +682,28 @@ class Group(Mapping):
                 refs = self._get_dim_scale_refs(name)
                 self._delete_dim_scale(name)
 
-        self._h5group.create_dataset(
+        h5ds = self._h5group.create_dataset(
             h5name,
             shape,
             dtype=dtype,
             data=data,
+            chunks=chunks,
             fillvalue=fillvalue,
             track_order=self._track_order,
             **kwargs,
         )
+
+        if chunks in {None, True} and (None in maxshape):
+            h5netcdf_chunks = _get_default_chunksizes(maxshape, dtype)
+            warnings.warn(
+                "Using h5py's default chunking with unlimited dimensions can lead "
+                "to increased file sizes and degraded performance (using chunks: %r). "
+                'Consider passing ``chunks="h5netcdf"`` (would give chunks: %r; '
+                "default in h5netcdf >= 1.0), or set chunk sizes explicitly. "
+                'To silence this warning, pass ``chunks="h5py"``. '
+                % (h5ds.chunks, h5netcdf_chunks),
+                FutureWarning,
+            )
 
         self._variables[h5name] = self._variable_cls(self, h5name, dimensions)
         variable = self._variables[h5name]
@@ -707,7 +730,14 @@ class Group(Mapping):
         return variable
 
     def create_variable(
-        self, name, dimensions=(), dtype=None, data=None, fillvalue=None, **kwargs
+        self,
+        name,
+        dimensions=(),
+        dtype=None,
+        data=None,
+        fillvalue=None,
+        chunks=None,
+        **kwargs,
     ):
         if name.startswith("/"):
             return self._root.create_variable(
@@ -718,7 +748,7 @@ class Group(Mapping):
         for k in keys[:-1]:
             group = group._require_child_group(k)
         return group._create_child_variable(
-            keys[-1], dimensions, dtype, data, fillvalue, **kwargs
+            keys[-1], dimensions, dtype, data, fillvalue, chunks, **kwargs
         )
 
     def _get_child(self, key):
@@ -1172,3 +1202,71 @@ class File(Group):
             self.mode,
         )
         return "\n".join([header] + self._repr_body())
+
+
+def _get_default_chunksizes(dimsizes, dtype):
+    # This is a modified version of h5py's default chunking heuristic
+    # https://github.com/h5py/h5py/blob/aa31f03bef99e5807d1d6381e36233325d944279/h5py/_hl/filters.py#L334-L389
+    # (published under BSD-3-Clause, included at licenses/H5PY_LICENSE.txt)
+    # See also https://github.com/h5py/h5py/issues/2029 for context.
+
+    CHUNK_BASE = 16 * 1024  # Multiplier by which chunks are adjusted
+    CHUNK_MIN = 8 * 1024  # Soft lower limit (8k)
+    CHUNK_MAX = 1024 * 1024  # Hard upper limit (1M)
+
+    type_size = np.dtype(dtype).itemsize
+
+    is_unlimited = np.array([x is None for x in dimsizes])
+
+    # For unlimited dimensions start with a guess of 1024
+    chunks = np.array([x if x is not None else 1024 for x in dimsizes], dtype="=f8")
+
+    ndims = len(dimsizes)
+    if ndims == 0:
+        raise ValueError("Chunks not allowed for scalar datasets.")
+
+    if not np.all(np.isfinite(chunks)):
+        raise ValueError("Illegal value in chunk tuple")
+
+    # Determine the optimal chunk size in bytes using a PyTables expression.
+    # This is kept as a float.
+    dset_size = np.product(chunks[~is_unlimited]) * type_size
+    target_size = CHUNK_BASE * (2 ** np.log10(dset_size / (1024 * 1024)))
+
+    if target_size > CHUNK_MAX:
+        target_size = CHUNK_MAX
+    elif target_size < CHUNK_MIN:
+        target_size = CHUNK_MIN
+
+    i = 0
+    while True:
+        # Repeatedly loop over the axes, dividing them by 2.
+        # Start by reducing unlimited axes first.
+        # Stop when:
+        # 1a. We're smaller than the target chunk size, OR
+        # 1b. We're within 50% of the target chunk size, AND
+        #  2. The chunk is smaller than the maximum chunk size
+
+        idx = i % ndims
+
+        chunk_bytes = np.product(chunks) * type_size
+
+        done = (
+            chunk_bytes < target_size
+            or abs(chunk_bytes - target_size) / target_size < 0.5
+        ) and chunk_bytes < CHUNK_MAX
+
+        if done:
+            break
+
+        if np.product(chunks) == 1:
+            break  # Element size larger than CHUNK_MAX
+
+        nelem_unlim = np.product(chunks[is_unlimited])
+
+        if nelem_unlim == 1 or is_unlimited[idx]:
+            chunks[idx] = np.ceil(chunks[idx] / 2.0)
+
+        i += 1
+
+    return tuple(int(x) for x in chunks)